hpsearch/requirements.txt
google-api-python-client==1.6.3
google-cloud-storage==1.4.0
kubernetes==3.0.0

scikit-learn==0.19.0
numpy==1.13.1
scipy==0.19.1
pyyaml==3.12
